backend/requirements.txt
aniso8601==10.0.1
blinker==1.9.0
cachelib==0.13.0
click==8.3.0
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
packaging==25.0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
